Create IncDec component, refactor index and cart to use it, base structure for cart changed
