+++ components/elements/CartIncDec/CartIncDec.module.scss
@@ -1,0 +1,29 @@
+@import "@/styles/main";
+.IncDec {
+  background-color: white;
+  margin: 0 auto;
+  display: flex;
+  max-width: 120px;
+  justify-content: space-between;
+  align-items: center;
+  height: 40px;
+  border-radius: 20px;
+  border: solid lightgray 1px;
+  button {
+    background-color: inherit;
+    border: 0;
+    width: 40px;
+    height: 38px;
+    border-radius: 50%;
+    font-weight: bold;
+    font-size: 1.6rem;
+    color: gray;
+    &:hover {
+      cursor: pointer;
+    }
+  }
+  span {
+    width: 50px;
+    text-align: center;
+  }
+}
