+++ styles/page-styles/Cart.module.scss
@@ -4,7 +4,7 @@
   display: flex;
 }
 .CTable {
-  width: calc(100% * (9 / 12));
+  width: 100%;
   table {
     width: 100%;
     border-collapse: separate;
@@ -12,7 +12,7 @@
   }
   td,
   th {
-    width: calc(100% * (1 / 5));
+    width: calc(100% * (1 / 6));
     vertical-align: middle;
     text-align: center;
     height: 70px;
@@ -25,8 +25,4 @@
   tr:nth-child(2) {
     background-color: #f9f9f9;
   }
-}
-.CSummary {
-  width: calc(100% * (3 / 12));
-  margin-left: 2rem;
 }
