+++ styles/page-styles/Home.module.scss
@@ -41,32 +41,6 @@
   display: flex;
   flex-direction: column;
   align-items: center;
-  &__quantity {
-    background-color: white;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    height: 40px;
-    border-radius: 20px;
-    border: solid lightgray 1px;
-    button {
-      background-color: inherit;
-      border: 0;
-      width: 40px;
-      height: 38px;
-      border-radius: 50%;
-      font-weight: bold;
-      font-size: 1.6rem;
-      color: gray;
-      &:hover {
-        cursor: pointer;
-      }
-    }
-    span {
-      width: 50px;
-      text-align: center;
-    }
-  }
   &__add {
     margin-top: 25px;
     width: 140px;
